mentee feedback messages and production fixes

+++ src/components/bootcamp.ts
@@ -1,5 +1,6 @@
 import { GuildMember, Role, TextChannel, Message, GuildChannel, VoiceState } from 'discord.js';
 import { CommandoClient } from 'discord.js-commando';
+import { getEmojiByName } from '../components/emojis';
 const Keyv = require('keyv');
 
 const BOOTCAMP_GUILD_ID: string = process.env.BOOTCAMP_GUILD_ID || '.';
@@ -66,23 +67,25 @@ export const controlMentorMenteeCalls = async (oldMember: VoiceState, newMember:
 
     if (chatChannel) {
       const mentorRole = await BootcampSettings.get('mentor_role');
-      if (leaver.roles.cache.map((role) => role.id).includes(mentorRole)) {
-        if (
+      if ((leaver.roles.cache.map((role) => role.id).includes(mentorRole)) && (
           oldUserChannel.members.filter((member: GuildMember) =>
             member.roles.cache.map((role) => role.id).includes(mentorRole)
           ).size === 0
-        ) {
-          try {
-            chatChannel.delete();
-            oldUserChannel.delete();
-          } catch (error) {
-            console.log('Channel already deleted');
-          }
+        )
+      ) {
+        try {
+          chatChannel.delete();
+          oldUserChannel.delete();
+        } catch (error) {
+          console.log('Channel already deleted');
         }
       } else {
         (async (): Promise<void> => {
           const fetched = await chatChannel.messages.fetch({ limit: 100 }).catch(console.log);
           if (fetched) chatChannel.bulkDelete(fetched);
+          if (!leaver.roles.cache.map((role) => role.id).includes(mentorRole)) {
+            leaver.send(`Thanks for taking part in Bootcamp: Resume Reviews! ${getEmojiByName('codeyLove')?.toString()}  I hope this event provided value for your future career prospects. ${getEmojiByName('codeyCoding2')?.toString()}  In order to improve future events, I'd love it if you filled out this feedback form: https://bit.ly/bootcamp-mentee-feedback ! Thanks so much ${getEmojiByName('codeyLove')?.toString()}`);
+          }
         })();
       }
       chatChannel?.updateOverwrite(leaver, {
